Extract jumptodate
